2009-03-02  Jeff Johnston  <jjohnstn@redhat.com>
        * libc/stdlib/wctomb_r.c (_wctomb_r): When checking single-byte
        charset, cast wchar to size_t in case wchar_t is signed.
        * libc/stdlib/wctomb.c (wctomb): Add similar single-byte check.

+++ newlib/libc/stdlib/wctomb_r.c
@@ -207,7 +207,7 @@ _DEFUN (_wctomb_r, (r, s, wchar, state),
     return 0;
  
   /* otherwise we are dealing with a single byte character */
-  if (wchar >= 0x100)
+  if ((size_t)wchar >= 0x100)
     {
       r->_errno = EILSEQ;
       return -1;
